tools/nest-gen/requirements.txt
openai>=1.30.0
edge-tts>=6.1.9
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
Pillow>=10.0.0
